Commands structure refactoring
