+++ calculator/README.md
@@ -95,19 +95,19 @@ cd frontend-console
 Then, send an input as follows:
 
 ```shell
-yarn start send --input "3 + 2 + 5 - 9"
+yarn start input send --payload "3 + 2 + 5 - 9"
 ```
 
 In order to verify the notices generated by your inputs, run the command:
 
 ```shell
-yarn start notices
+yarn start notice list
 ```
 
 The response should be something like this:
 
 ```json
-[{"epoch": "0", "input": "1", "notice": "0", "payload": "1"}]
+[{ "epoch": "0", "input": "1", "notice": "0", "payload": "1" }]
 ```
 
 ## Running the environment in host mode
