+++ converter/README.md
@@ -6,7 +6,7 @@ The DApp receives a JSON string as input, containing both the string to be conve
 This DApp receives as input a JSON string in the following format:
 
 ```json
-{"transform": "upper", "message": "hello from cartesi"}
+{ "transform": "upper", "message": "hello from cartesi" }
 ```
 
 Where `message` indicates the text to be converted, and `transform` specifies the operation to be performed on it.
@@ -26,19 +26,19 @@ cd frontend-console
 Then, send an input as follows:
 
 ```shell
-yarn start send --input '{"transform": "upper", "message": "hello from cartesi"}'
+yarn start input send --payload '{"transform": "upper", "message": "hello from cartesi"}'
 ```
 
 In order to verify the notices generated by your inputs, run the command:
 
 ```shell
-yarn start notices
+yarn start notice list
 ```
 
 The response should be something like this:
 
 ```json
-[{"epoch": "0", "input": "1", "notice": "0", "payload": "HELLO FROM CARTESI"}]
+[{ "epoch": "0", "input": "1", "notice": "0", "payload": "HELLO FROM CARTESI" }]
 ```
 
 ## Running the back-end in host mode
