+++ custom-dapps/template/README.md
@@ -38,13 +38,13 @@ Ensure that the [application has already been built](https://github.com/cartesi/
 From within the `frontend-console` directory, you can send an input as follows:
 
 ```shell
-yarn start send --input "Hello there"
+yarn start input send --payload "Hello there"
 ```
 
 In order to verify the notices generated by your inputs, run the command:
 
 ```shell
-yarn start notices
+yarn start notice list
 ```
 
 The response should be something like this:
@@ -123,13 +123,13 @@ export RPC_URL=<https://your.rpc.gateway>
 Then, inputs can be sent by specifying the DApp contract's address, as follows:
 
 ```shell
-yarn start send --input "Hello there" --addressFile path/to/template/deployments/<network>/template.address
+yarn start input send --payload "Hello there" --addressFile path/to/template/deployments/<network>/template.address
 ```
 
 Resulting notices can then be retrieved by querying the local Cartesi Node, as before:
 
 ```shell
-yarn start notices
+yarn start notice list
 ```
 
 ## Running the back-end in host mode
