+++ echo-js/README.md
@@ -16,13 +16,13 @@ cd frontend-console
 Then, send an input as follows:
 
 ```shell
-yarn start send --input "Hello there"
+yarn start input send --payload "Hello there"
 ```
 
 In order to verify the notices generated by your inputs, run the command:
 
 ```shell
-yarn start notices
+yarn start notice list
 ```
 
 The response should be something like this:
