+++ erc20/README.md
@@ -35,13 +35,20 @@ yarn start erc20 deposit --amount 100
 In order to verify the notices generated by your inputs, run the command:
 
 ```shell
-yarn start notices
+yarn start notice list
 ```
 
 The response should be something like this:
 
 ```json
-[{"epoch":"0","input":"1","notice":"0","payload":"Deposit received from: 0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266; ERC-20: 0xa513e6e4b8f2a923d98304ec87f64353c4d5c853; Amount: 100"}]
+[
+  {
+    "epoch": "0",
+    "input": "1",
+    "notice": "0",
+    "payload": "Deposit received from: 0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266; ERC-20: 0xa513e6e4b8f2a923d98304ec87f64353c4d5c853; Amount: 100"
+  }
+]
 ```
 
 ## Running the environment in host mode
