+++ frontend-console/src/commands/input.ts
@@ -1,0 +1,20 @@
+// Copyright 2022 Cartesi Pte. Ltd.
+
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not use
+// this file except in compliance with the License. You may obtain a copy of the
+// License at http://www.apache.org/licenses/LICENSE-2.0
+
+// Unless required by applicable law or agreed to in writing, software distributed
+// under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+// CONDITIONS OF ANY KIND, either express or implied. See the License for the
+// specific language governing permissions and limitations under the License.
+
+import { Argv } from "yargs";
+
+export const command = "input <command>";
+export const desc = "Operations with inputs";
+
+export const builder = (yargs: Argv) =>
+    yargs.commandDir("input_commands", { extensions: ["js", "ts"] });
+
+export const handler = () => {};
