+++ knn/README.md
@@ -4,7 +4,7 @@ This example implements a simple Machine Learning classification algorithm withi
 
 In this DApp, we use the [k-Nearest Neighbor algorithm](https://en.wikipedia.org/wiki/K-nearest_neighbors_algorithm) to predict the class of a given input based on a previously known dataset of samples.
 
-We use the classic [Iris flower dataset](https://en.wikipedia.org/wiki/Iris_flower_data_set), which contains 3 classes of 50 instances each, where each class refers to a species of the [Iris flower](https://en.wikipedia.org/wiki/Iris_(plant)).
+We use the classic [Iris flower dataset](https://en.wikipedia.org/wiki/Iris_flower_data_set), which contains 3 classes of 50 instances each, where each class refers to a species of the [Iris flower](<https://en.wikipedia.org/wiki/Iris_(plant)>).
 One class is linearly separable from the other two, however the latter are NOT linearly separable from each other.
 The actual data used in this example is available [here](https://archive.ics.uci.edu/ml/datasets/iris).
 
@@ -27,7 +27,7 @@ Once started, the DApp will receive input samples and predict its classification
 Each input should be given as a JSON string, such as the following:
 
 ```json
-{"sl": 4.9, "sw": 3.0, "pl": 1.4, "pw": 0.3}
+{ "sl": 4.9, "sw": 3.0, "pl": 1.4, "pw": 0.3 }
 ```
 
 Where `sl` represents the sepal length, `sw` is the sepal width, `pl` is the petal length and `pw` is the petal width.
@@ -46,19 +46,19 @@ cd frontend-console
 Then, send an input as follows:
 
 ```shell
-yarn start send --input '{"sl": 4.9, "sw": 3.0, "pl": 1.4, "pw": 0.3}'
+yarn start input send --payload '{"sl": 4.9, "sw": 3.0, "pl": 1.4, "pw": 0.3}'
 ```
 
 In order to verify the notices generated by your inputs, run the command:
 
 ```shell
-yarn start notices
+yarn start notice list
 ```
 
 The response should be something like this:
 
 ```json
-[{"epoch":"0","input":"1","notice":"0","payload":"Iris-setosa"}]
+[{ "epoch": "0", "input": "1", "notice": "0", "payload": "Iris-setosa" }]
 ```
 
 The payload corresponds to the Iris flower species class predicted by the k-NN algorithm.
