+++ m2cgen/README.md
@@ -17,7 +17,7 @@ The model currently takes into account only three characteristics of a person to
 As such, inputs to the DApp should be given as a JSON string such as the following:
 
 ```json
-{"Age": 37, "Sex": "male", "Embarked": "S"}
+{ "Age": 37, "Sex": "male", "Embarked": "S" }
 ```
 
 The predicted classification result will be given as `0` (did not survive) or `1` (did survive).
@@ -36,19 +36,19 @@ cd frontend-console
 Then, send an input as follows:
 
 ```shell
-yarn start send --input '{"Age": 37, "Sex": "male", "Embarked": "S"}'
+yarn start input send --payload '{"Age": 37, "Sex": "male", "Embarked": "S"}'
 ```
 
 In order to verify the notices generated by your inputs, run the command:
 
 ```shell
-yarn start notices
+yarn start notice list
 ```
 
 The response should be something like this:
 
 ```json
-[{"epoch":"0","input":"1","notice":"0","payload":"0"}]
+[{ "epoch": "0", "input": "1", "notice": "0", "payload": "0" }]
 ```
 
 The payload corresponds to whether the person survived (`1`) or not (`0`).
