+++ sqlite/README.md
@@ -18,19 +18,19 @@ cd frontend-console
 Then, send an input to create a table as follows:
 
 ```shell
-yarn start send --input "CREATE TABLE Persons (name text, age int)"
+yarn start input send --payload "CREATE TABLE Persons (name text, age int)"
 ```
 
 Next, add an entry to the newly created table by submitting an SQL `INSERT` statement as an input:
 
 ```shell
-yarn start send --input "INSERT INTO Persons VALUES ('Peter', 32)"
+yarn start input send --payload "INSERT INTO Persons VALUES ('Peter', 32)"
 ```
 
 Once data has been inserted into the database, it can be queried using a regular SQL `SELECT` statement such as the following:
 
 ```shell
-yarn start send --input "SELECT * FROM Persons"
+yarn start input send --payload "SELECT * FROM Persons"
 ```
 
 Note that the query's results will not be retrieved immediately. Rather, whenever a submitted input corresponds to a valid SQL query, the DApp will make the corresponding results available in the form of a _notice_.
@@ -38,13 +38,13 @@ Note that the query's results will not be retrieved immediately. Rather, wheneve
 In order to verify the notices generated by your inputs, run the command:
 
 ```shell
-yarn start notices
+yarn start notice list
 ```
 
 The response should be something like this:
 
 ```json
-[{"epoch":"0","input":"3","notice":"0","payload":"[[\"Peter\", 32]]"}]
+[{ "epoch": "0", "input": "3", "notice": "0", "payload": "[[\"Peter\", 32]]" }]
 ```
 
 ## Running the environment in host mode
